+++ src/Bootstrap.php
@@ -2,7 +2,8 @@
 
 namespace webhubworks\flare;
 
-use yii\base\Application;
+use craft\console\Application as ConsoleApplication;
+use craft\web\Application as WebApplication;
 use yii\base\BootstrapInterface;
 
 class Bootstrap implements BootstrapInterface
@@ -11,7 +12,7 @@ class Bootstrap implements BootstrapInterface
      * Installs our components during the bootstrap process to get us loaded
      * sooner in case something crashes.
      *
-     * @param Application|\craft\web\Application $app
+     * @param WebApplication|ConsoleApplication $app
      */
     public function bootstrap($app): void
     {